+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed: inconsistent polyline connectors with endpoints into transformed
+  groups.
+
 ## [0.26.0 - 2026-01-03]
 
 - Fixed: location references now respect transforms which may have moved